Beaver contact (#11)
* added mason's video to html file
git push

* added Beth's contact page and link to the navigation
The contact a href is in the index file in the navigation and acts as should next to the login button

* Index.html progress

In this commit, I have designed the homepage and added images to the different sections.  I have included the Bungee fonts to the folders as well.  Lastly, I cleaned up the working project folder to to include only what is being used

* revert